feat: add CodeBlock extension with lowlight support and update related imports, improve bubble menu code quality
