+++ apps/web/package.json
@@ -49,8 +49,8 @@
     "fumadocs-mdx": "^15.0.12",
     "fumadocs-ui": "^16.10.2",
     "jotai": "^2.19.1",
-    "lowlight": "^3.3.0",
     "lru-cache": "^11.5.2",
+    "lowlight": "^3.3.0",
     "lucide-react": "catalog:",
     "motion": "^12.38.0",
     "next": "^16.2.0",
