+++ apps/web/src/app/layout.tsx
@@ -7,7 +7,8 @@ import { fontVariables } from "@/lib/fonts";
 import { cn } from "@/lib/utils";
 import { SoftwareApplicationJsonLd, WebSiteJsonLd } from "@/seo/json-ld";
 
-import "@/index.css";
+import "@editorcn/ui/globals.css";
+import "@/styles/themes.css";
 
 export const metadata: Metadata = {
   title: {
