+++ apps/web/src/styles/themes.css
@@ -1,4 +1,4 @@
-.theme-default .theme-container {
+/* .theme-default .theme-container {
   --chart-1: var(--color-blue-300);
   --chart-2: var(--color-blue-500);
   --chart-3: var(--color-blue-600);
@@ -332,4 +332,4 @@
     --sidebar-primary-foreground: var(--color-violet-50);
     --sidebar-ring: var(--color-violet-900);
   }
-}
+} */
