+++ packages/block-editor/package.json
@@ -28,22 +28,25 @@
   "peerDependencies": {
     "@base-ui/react": "^1.0.0",
     "@tiptap/core": ">=2.11.5 <4",
-    "@tiptap/extension-code-block-lowlight": ">=2.11.5 <4",
     "@tiptap/extension-drag-handle-react": ">=2.11.5 <4",
     "@tiptap/extension-link": ">=2.11.5 <4",
     "@tiptap/extension-placeholder": ">=2.11.5 <4",
     "@tiptap/extension-underline": ">=2.11.5 <4",
-    "@tiptap/suggestion": ">=2.11.5 <4",
     "@tiptap/pm": ">=2.11.5 <4",
     "@tiptap/react": ">=2.11.5 <4",
     "@tiptap/starter-kit": ">=2.11.5 <4",
+    "@tiptap/suggestion": ">=2.11.5 <4",
     "class-variance-authority": "^0.7.1",
-    "lowlight": "^3.3.0",
     "react": "^18 || ^19",
-    "react-dom": "^18 || ^19"
+    "react-dom": "^18 || ^19",
+    "@tiptap/extension-code-block-lowlight": ">=2.11.5 <4",
+    "lowlight": ">=3.0.0 <4"
   },
   "peerDependenciesMeta": {
     "@tiptap/extension-code-block-lowlight": {
+      "optional": true
+    },
+    "lowlight": {
       "optional": true
     },
     "@tiptap/extension-placeholder": {
