+++ packages/block-editor/src/icons.tsx
@@ -57,7 +57,169 @@ export interface BlockEditorIcons {
   copyIcon: React.ReactNode;
   deleteIcon: React.ReactNode;
   dropdownArrowIcon: React.ReactNode;
+  codeBlockLanguageIcon: React.ReactNode;
+  languageIcons: Record<string, React.ReactNode>;
 }
+
+export const DEFAULT_LANGUAGE_ICONS: Record<string, React.ReactNode> = {
+  javascript: (
+    <Svg>
+      <rect x="2" y="2" width="20" height="20" rx="2" />
+      <path d="M10 16V8h1.5c1.4 0 2.5.6 2.5 2 0 1-.5 1.5-1.2 1.8C13.8 12.2 15 13 15 15c0 1.8-1.4 3-3 3H9" />
+      <path d="M11 8h2" />
+    </Svg>
+  ),
+  typescript: (
+    <Svg>
+      <rect x="2" y="2" width="20" height="20" rx="2" />
+      <path d="M8 16V8h3c1.7 0 3 1.1 3 2.8 0 1.5-.8 2.5-2 2.8.9.3 1.5 1.1 1.5 2.2 0 1.7-1.3 3.2-3 3.2H8" />
+      <path d="M9 8h3" />
+    </Svg>
+  ),
+  html: (
+    <Svg>
+      <path d="M4 3l1.5 18L12 23l6.5-2L20 3z" />
+      <path d="M8 8h8l-.5 5-3.5 1-3.5-1z" />
+      <path d="M12 14v6" />
+    </Svg>
+  ),
+  css: (
+    <Svg>
+      <path d="M4 3l1.5 18L12 23l6.5-2L20 3z" />
+      <path d="M8 8h8l-.3 3-3.7 1.2L7.2 11z" />
+      <circle cx="12" cy="16" r="1.5" />
+      <path d="M9.5 19.5l-.5-3 3 1 3-1-.5 3" />
+    </Svg>
+  ),
+  json: (
+    <Svg>
+      <path d="M8 3H7a2 2 0 0 0-2 2v4a2 2 0 0 1-2 2 2 2 0 0 1 2 2v4a2 2 0 0 0 2 2" />
+      <path d="M16 3h1a2 2 0 0 1 2 2v4a2 2 0 0 0 2 2 2 2 0 0 0-2 2v4a2 2 0 0 1-2 2" />
+    </Svg>
+  ),
+  python: (
+    <Svg>
+      <path d="M11 2c-2.2 0-4 .9-4 3v2h4v1H6.5C4 9 2 10.8 2 13.5S4 18 6.5 18H8v-2.5C8 14 9 13 10.5 13h4c1.5 0 2.5-1 2.5-2.5v-6C17 3 14 2 11 2" />
+      <circle cx="8.5" cy="5.5" r=".8" fill="currentColor" stroke="none" />
+      <path d="M13 22c2.2 0 4-.9 4-3v-2h-4v-1h4.5c2.5 0 4.5-1.8 4.5-4.5S19.5 7 17 7H16v2.5c0 1.5-1 2.5-2.5 2.5h-4c-1.5 0-2.5 1-2.5 2.5v6C7 21 10 22 13 22" />
+      <circle cx="15.5" cy="18.5" r=".8" fill="currentColor" stroke="none" />
+    </Svg>
+  ),
+  rust: (
+    <Svg>
+      <circle cx="12" cy="12" r="9.5" />
+      <circle cx="12" cy="12" r="2.5" />
+      <path d="M12 2.5v3M12 18.5v3M2.5 12h3M18.5 12h3" />
+      <path d="M5.6 5.6l2.1 2.1M16.3 16.3l2.1 2.1M5.6 18.4l2.1-2.1M16.3 7.7l2.1-2.1" />
+    </Svg>
+  ),
+  go: (
+    <Svg>
+      <path d="M3 12h3m12 0h3" />
+      <path d="M12 3v4m0 10v4" />
+      <circle cx="12" cy="12" r="3.5" />
+      <path d="M12 8.5V7" />
+      <path d="M12 17v-1.5" />
+    </Svg>
+  ),
+  java: (
+    <Svg>
+      <path d="M8 4c0 0-4 1.5-4 7s4 7 4 7" />
+      <path d="M16 4c0 0 4 1.5 4 7s-4 7-4 7" />
+      <ellipse cx="12" cy="14" rx="4.5" ry="5" />
+      <path d="M9.5 17.5L8.5 21" />
+      <path d="M14.5 17.5L15.5 21" />
+    </Svg>
+  ),
+  c: (
+    <Svg>
+      <circle cx="12" cy="12" r="9" />
+      <path d="M15 8.5c-1-1-2.5-1.5-4-.5s-2 3-.5 4.5l4 4c1.5 1.5 3.5.5 4-.5" />
+    </Svg>
+  ),
+  cpp: (
+    <Svg>
+      <circle cx="10" cy="12" r="8" />
+      <path d="M12.5 9c-1-1-2.5-1-3 0s-1 2.5 0 3.5l3 3c1 1 2.5 1 3 0" />
+      <circle cx="16.5" cy="15" r=".8" fill="currentColor" stroke="none" />
+      <circle cx="18.5" cy="12" r=".8" fill="currentColor" stroke="none" />
+      <circle cx="16.5" cy="9" r=".8" fill="currentColor" stroke="none" />
+    </Svg>
+  ),
+  ruby: (
+    <Svg>
+      <polygon points="12 2 20 8 12 22 4 8" />
+      <line x1="4" y1="8" x2="20" y2="8" />
+      <line x1="12" y1="2" x2="12" y2="22" />
+    </Svg>
+  ),
+  php: (
+    <Svg>
+      <ellipse cx="12" cy="12" rx="9" ry="6" />
+      <path d="M9.5 7v10" />
+      <path d="M9.5 7c2.5-1.5 5.5-.5 6.5 1.5s.5 5-1.5 6.5-5.5 1-7 0" />
+    </Svg>
+  ),
+  swift: (
+    <Svg>
+      <path d="M18 4c-4 6-6.5 8.5-8.5 11 2.5-.5 4.5-1.5 5.5-4-1.5 3.5-4 6-7 7 4 1 7.5-.5 9.5-3 .5 3-1 5.5-3 6.5 5-1.5 8-5 9-9.5-1.5 4-4 6.5-6.5 8" />
+    </Svg>
+  ),
+  kotlin: (
+    <Svg>
+      <rect x="3" y="3" width="18" height="18" rx="2" />
+      <path d="M8 7v10" />
+      <path d="M8 12l5.5-5.5" />
+      <path d="M13.5 12H17" />
+    </Svg>
+  ),
+  sql: (
+    <Svg>
+      <ellipse cx="12" cy="6" rx="8" ry="3" />
+      <path d="M4 6v6c0 1.7 3.6 3 8 3s8-1.3 8-3V6" />
+      <path d="M4 12v6c0 1.7 3.6 3 8 3s8-1.3 8-3v-6" />
+    </Svg>
+  ),
+  bash: (
+    <Svg>
+      <rect x="2" y="3" width="20" height="18" rx="2" />
+      <path d="M6 8.5l3.5 3.5L6 15.5" />
+      <path d="M12 15.5h6" />
+    </Svg>
+  ),
+  markdown: (
+    <Svg>
+      <rect x="2" y="4" width="20" height="16" rx="2" />
+      <path d="M6 16V8l3 4 3-4v8" />
+      <path d="M18 12l-3-4-3 4" />
+      <path d="M15 16V8" />
+    </Svg>
+  ),
+  yaml: (
+    <Svg>
+      <rect x="2" y="3" width="20" height="18" rx="2" />
+      <path d="M7 8h3" />
+      <path d="M7 12h5" />
+      <path d="M7 16h4" />
+    </Svg>
+  ),
+  xml: (
+    <Svg>
+      <polyline points="4 7 2 12 4 17" />
+      <polyline points="20 7 22 12 20 17" />
+      <line x1="14" y1="4" x2="10" y2="20" />
+      <line x1="7" y1="8" x2="17" y2="16" />
+    </Svg>
+  ),
+  plaintext: (
+    <Svg>
+      <rect x="3" y="2" width="18" height="20" rx="2" />
+      <line x1="7" y1="7" x2="17" y2="7" />
+      <line x1="7" y1="11" x2="15" y2="11" />
+      <line x1="7" y1="15" x2="13" y2="15" />
+    </Svg>
+  ),
+};
 
 export const DEFAULT_ICONS: BlockEditorIcons = {
   slashTextIcon: (
@@ -249,4 +411,11 @@ export const DEFAULT_ICONS: BlockEditorIcons = {
       <path d="m6 9 6 6 6-6" />
     </Svg>
   ),
+  codeBlockLanguageIcon: (
+    <Svg>
+      <polyline points="16 18 22 12 16 6" />
+      <polyline points="8 6 2 12 8 18" />
+    </Svg>
+  ),
+  languageIcons: DEFAULT_LANGUAGE_ICONS,
 };
