+++ packages/block-editor/src/index.ts
@@ -1,9 +1,10 @@
 export { BlockEditor } from './block-editor';
 export { useBlockEditorContext } from './context';
-export { BubbleMenu } from './bubble-menu';
+export { BubbleMenu } from './bubble-menu/index';
 export { SlashCommand, getSlashCommandSuggestion, defaultSlashCommandItems } from './extensions';
+export { CodeBlock } from './extensions/code-block';
 export type { SlashCommandSuggestionItem, OnCommandSelect } from './extensions';
 export type { BlockEditorProps, BlockEditorLabels } from './types';
 export { DEFAULT_BLOCK_EDITOR_LABELS } from './labels';
 export type { BlockEditorIcons } from './icons';
-export { DEFAULT_ICONS } from './icons';
+export { DEFAULT_ICONS, DEFAULT_LANGUAGE_ICONS } from './icons';
