+++ packages/block-editor/src/style.css
@@ -124,7 +124,24 @@
 
 .block-editor-bubble-btn--active {
   color: var(--primary);
-  background: var(--muted);
+  background: var(--accent);
+}
+
+.block-editor-copy-icon {
+  display: inline-flex;
+  width: 1rem;
+  height: 1rem;
+  transition: transform 0.2s cubic-bezier(0.34, 1.56, 0.64, 1), color 0.15s ease;
+}
+
+.block-editor-copy-icon > svg {
+  width: 1rem;
+  height: 1rem;
+}
+
+.block-editor-copy-icon--copied {
+  color: var(--primary);
+  transform: scale(1.15);
 }
 
 .block-editor-bubble-btn-text {
@@ -165,6 +182,13 @@
 
 .block-editor-bubble-dropdown--align {
   min-width: 130px;
+}
+
+.block-editor-bubble-dropdown--language {
+  min-width: 160px;
+  max-height: 320px;
+  overflow-y: auto;
+  scrollbar-width: thin;
 }
 
 .block-editor-bubble-dropdown-item {
@@ -313,6 +337,68 @@
   background: none;
   padding: 0;
   font-weight: 400;
+}
+
+/* ── Syntax Highlighting (lowlight / highlight.js) ── */
+.block-editor-content .ProseMirror pre .hljs-comment,
+.block-editor-content .ProseMirror pre .hljs-quote {
+  color: #6c7086;
+}
+
+.block-editor-content .ProseMirror pre .hljs-variable,
+.block-editor-content .ProseMirror pre .hljs-template-variable,
+.block-editor-content .ProseMirror pre .hljs-attribute,
+.block-editor-content .ProseMirror pre .hljs-tag,
+.block-editor-content .ProseMirror pre .hljs-regexp,
+.block-editor-content .ProseMirror pre .hljs-link,
+.block-editor-content .ProseMirror pre .hljs-name,
+.block-editor-content .ProseMirror pre .hljs-selector-id,
+.block-editor-content .ProseMirror pre .hljs-selector-class {
+  color: #89b4fa;
+}
+
+.block-editor-content .ProseMirror pre .hljs-number,
+.block-editor-content .ProseMirror pre .hljs-meta,
+.block-editor-content .ProseMirror pre .hljs-built_in,
+.block-editor-content .ProseMirror pre .hljs-builtin-name,
+.block-editor-content .ProseMirror pre .hljs-literal,
+.block-editor-content .ProseMirror pre .hljs-type,
+.block-editor-content .ProseMirror pre .hljs-params {
+  color: #fab387;
+}
+
+.block-editor-content .ProseMirror pre .hljs-string,
+.block-editor-content .ProseMirror pre .hljs-symbol,
+.block-editor-content .ProseMirror pre .hljs-bullet {
+  color: #a6e3a1;
+}
+
+.block-editor-content .ProseMirror pre .hljs-title,
+.block-editor-content .ProseMirror pre .hljs-section {
+  color: #f9e2af;
+}
+
+.block-editor-content .ProseMirror pre .hljs-keyword,
+.block-editor-content .ProseMirror pre .hljs-selector-tag {
+  color: #cba6f7;
+}
+
+.block-editor-content .ProseMirror pre .hljs-emphasis {
+  font-style: italic;
+}
+
+.block-editor-content .ProseMirror pre .hljs-strong {
+  font-weight: 700;
+}
+
+.block-editor-content .ProseMirror pre .hljs-addition {
+  color: #a6e3a1;
+  background-color: color-mix(in srgb, #a6e3a1 10%, transparent);
+}
+
+.block-editor-content .ProseMirror pre .hljs-deletion {
+  color: #f38ba8;
+  background-color: color-mix(in srgb, #f38ba8 10%, transparent);
 }
 
 .block-editor-content .ProseMirror code {
