+++ packages/editor/src/extensions/code-block.ts
@@ -1,0 +1,9 @@
+import LowlightCodeBlock from "@tiptap/extension-code-block-lowlight";
+import { common, createLowlight } from "lowlight";
+
+const lowlight = createLowlight(common);
+
+export const CodeBlock = LowlightCodeBlock.configure({
+  lowlight,
+  defaultLanguage: "javascript",
+});
