+++ packages/editor/src/index.ts
@@ -1,12 +1,14 @@
 export { RichTextEditor } from "./rte-text-editor";
 export { useRichTextEditorContext } from "./rte-context";
+export { BubbleMenu } from "./bubble-menu/index";
 export type { RichTextEditorProps } from "./types";
 export type { RichTextEditorVariant } from "./types";
 export type { RichTextEditorFooterProps, WordCountInfo } from "./types";
 export type { RichTextEditorLabels } from "./labels";
 export { DEFAULT_LABELS } from "./labels";
 export type { RichTextEditorIcons } from "./icons";
-export { DEFAULT_ICONS } from "./icons";
+export { DEFAULT_ICONS, DEFAULT_LANGUAGE_ICONS } from "./icons";
 export { Link } from "./extensions";
+export { CodeBlock } from "./extensions/code-block";
 export { TwitterEmbed } from "./controls/rte-twitter-control";
 export { YouTubeEmbed } from "./controls/rte-youtube-control";
