+++ packages/editor/src/style.css
@@ -731,3 +731,161 @@ button.rte-control-button:active {
 .rte-resize-handle--corner:hover {
   transform: scale(1.15);
 }
+
+/* ── Bubble Menu ── */
+
+@keyframes rte-bubble-pop-in {
+  0% { opacity: 0; transform: scale(0.96) translateY(4px); }
+  100% { opacity: 1; transform: scale(1) translateY(0); }
+}
+
+.rte-bubble-menu {
+  display: flex;
+  align-items: center;
+  height: 2.25rem;
+  background: var(--popover);
+  border: 1px solid var(--border);
+  border-radius: var(--radius);
+  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.12);
+  padding: 0 0.25rem;
+  animation: rte-bubble-pop-in 0.2s cubic-bezier(0.16, 1, 0.3, 1) forwards;
+}
+
+.rte-bubble-group {
+  display: flex;
+  align-items: center;
+  gap: 1px;
+}
+
+.rte-bubble-separator {
+  width: 1px;
+  height: 1.25rem;
+  background: var(--border);
+  margin: 0 0.125rem;
+  flex-shrink: 0;
+}
+
+.rte-bubble-btn {
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+  height: 1.75rem;
+  padding: 0 0.375rem;
+  border: none;
+  border-radius: calc(var(--radius) - 4px);
+  background: transparent;
+  color: var(--popover-foreground);
+  cursor: pointer;
+  font-size: 0.8125rem;
+  white-space: nowrap;
+  transition: background-color 0.15s ease, color 0.15s ease;
+}
+
+.rte-bubble-btn:hover {
+  background: var(--accent);
+  color: var(--accent-foreground);
+}
+
+.rte-bubble-btn:active {
+  transform: scale(0.94);
+}
+
+.rte-bubble-btn--active {
+  color: var(--primary);
+  background: var(--accent);
+}
+
+.rte-bubble-btn-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  width: 1rem;
+  height: 1rem;
+}
+
+.rte-bubble-btn-icon svg {
+  width: 100%;
+  height: 100%;
+}
+
+.rte-bubble-btn-text {
+  max-width: 5rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.rte-bubble-overlay {
+  position: fixed;
+  inset: 0;
+  z-index: 999;
+}
+
+.rte-bubble-dropdown {
+  position: absolute;
+  top: 100%;
+  left: 0;
+  z-index: 1000;
+  min-width: 150px;
+  background: var(--popover);
+  border: 1px solid var(--border);
+  border-radius: var(--radius);
+  box-shadow: 0 8px 24px rgba(0, 0, 0, 0.15);
+  padding: 0.25rem;
+  margin-top: 4px;
+  transform-origin: top left;
+  animation: rte-bubble-pop-in 0.15s cubic-bezier(0.16, 1, 0.3, 1) forwards;
+}
+
+.rte-bubble-dropdown--language {
+  min-width: 160px;
+  max-height: 320px;
+  overflow-y: auto;
+  scrollbar-width: thin;
+}
+
+.rte-bubble-dropdown-item {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  width: 100%;
+  padding: 0.375rem 0.5rem;
+  font-size: 0.8125rem;
+  color: var(--popover-foreground);
+  background: transparent;
+  border: none;
+  border-radius: calc(var(--radius) - 4px);
+  cursor: pointer;
+  text-align: left;
+  transition: background-color 0.15s ease, color 0.15s ease;
+}
+
+.rte-bubble-dropdown-item:hover {
+  background: var(--accent);
+  color: var(--accent-foreground);
+}
+
+.rte-bubble-dropdown-item--active {
+  color: var(--primary);
+  background-color: color-mix(in srgb, var(--primary) 10%, transparent);
+}
+
+.rte-bubble-dropdown-icon {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  flex-shrink: 0;
+  color: var(--muted-foreground);
+  width: 1rem;
+  height: 1rem;
+  margin-left: auto;
+}
+
+.rte-bubble-dropdown-icon svg {
+  width: 100%;
+  height: 100%;
+}
+
+.rte-bubble-dropdown-item--active .rte-bubble-dropdown-icon {
+  color: var(--primary);
+}
