+++ packages/editor/tsconfig.json
@@ -11,4 +11,3 @@
     "include": ["src/**/*.ts", "src/**/*.tsx"],
     "exclude": ["node_modules"]
   }
-
